backend/requirements.txt
Flask==1.0.2
uWSGI==2.0.17.1
flask-cors==3.0.8
pyjwt==1.7.1
redis==3.3.11
flask-httpauth==3.3.0
pymongo==3.10.0
dnspython==1.16.0
Flask-GraphQL==2.0.1
graphene-mongo==0.2.8